Moved effects outside of the attack package and added Javadocs to them all.
