+++ source/core/src/main/com/csse3200/game/components/effects/EffectType.java
@@ -1,0 +1,7 @@
+package com.csse3200.game.components.effects;
+
+public enum EffectType {
+    STUN,
+    POISON,
+    KNOCKBACK
+}
