--- source/core/src/main/com/csse3200/game/components/npc/attack/attackeffects/Effect.java
@@ -1,12 +1,0 @@
-package com.csse3200.game.components.npc.attack.attackeffects;
-
-import com.csse3200.game.entities.Entity;
-
-/**
- * Interface for effects that can be applied to an entity.
- */
-public interface Effect {
-    void apply(Entity target);
-    void update(Entity target, float deltaTime);
-    void remove(Entity target);
-}
