+++ source/core/src/test/com/csse3200/game/components/npc/attack/attackeffects/KnockbackEffectsTest.java
@@ -2,6 +2,7 @@ package com.csse3200.game.components.npc.attack.attackeffects;
 
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.physics.box2d.Body;
+import com.csse3200.game.components.effects.KnockbackEffect;
 import com.csse3200.game.entities.Entity;
 import com.csse3200.game.extensions.GameExtension;
 import com.csse3200.game.physics.components.PhysicsComponent;
